requirements.txt
datasets
fastapi
langchain
langchain-community
langgraph
matplotlib
numpy
pandas
peft
pymilvus
pyyaml
requests
torch
transformers
unsloth
uvicorn
